Implementado a iniciação de dependencias, descoplado os métodos de tratamente e a interface do user_domain, adição de logs no controller, service e repository
