+++ src/model/user_domain.go
@@ -1,10 +1,5 @@
 package model
 
-import (
-	"crypto/md5"
-	"encoding/hex"
-)
-
 type userDomain struct {
 	id       string
 	email    string
@@ -13,38 +8,12 @@ type userDomain struct {
 	age      int8
 }
 
-type UserDomainInterface interface {
-	GetEmail() string
-	GetPassword() string
-	GetAge() int8
-	GetName() string
-	GetID() string
-
-	SetID(string)
-	EncryptPassword()
-}
-
-func NewUserDomain(email, password, name string, age int8) UserDomainInterface {
-	return &userDomain{
-		email:    email,
-		password: password,
-		name:     name,
-		age:      age}
-}
-
 func (ud *userDomain) SetID(id string) {
 	ud.id = id
 }
 
 func (ud *userDomain) GetID() string {
 	return ud.id
-}
-
-func (ud *userDomain) EncryptPassword() {
-	hash := md5.New()
-	defer hash.Reset()
-	hash.Write([]byte(ud.password))
-	ud.password = hex.EncodeToString(hash.Sum(nil))
 }
 
 func (ud userDomain) GetEmail() string {
