+++ src/model/user_domain_interface.go
@@ -1,0 +1,20 @@
+package model
+
+type UserDomainInterface interface {
+	GetEmail() string
+	GetPassword() string
+	GetAge() int8
+	GetName() string
+	GetID() string
+
+	SetID(string)
+	EncryptPassword()
+}
+
+func NewUserDomain(email, password, name string, age int8) UserDomainInterface {
+	return &userDomain{
+		email:    email,
+		password: password,
+		name:     name,
+		age:      age}
+}
